Add request and response struct version numbers and response timeout
We need to be able to detect a version mismatch between client and
server and the easiest way to do this is include a version number in
each structure used as a message.

Also, we don't want to wait forever for a response that is not coming,
use epoll to wait on server responses with a timeout.

Signed-off-by: Eric B Munson <eric@munsonfam.org>